CFU refractoring
